+++ src/React/EventLoop/LibEventLoop.php
@@ -167,7 +167,7 @@ class LibEventLoop implements LoopInterface
     /**
      * {@inheritdoc}
      */
-    public function nextTick(callable $listener)
+    public function onNextTick(callable $listener)
     {
         $this->nextTickQueue->add($listener);
     }