+++ CHANGELOG.md
@@ -7,6 +7,10 @@ CHANGELOG
   * BC break: Bump minimum PHP version to PHP 5.4, remove 5.3 specific hacks
   * Bug fix: [Dns] Properly resolve CNAME aliases
 
+  * Bug fix: [EventLoop] No error on removing non-existent streams (@clue)
+  * Bug fix: [EventLoop] Do not silently remove feof listeners in `LibEvLoop`
+  * Bug fix: [Stream] Correctly detect closed connections
+
 ### 0.3.2 (2013-05-10)
 
   * Feature: [Dns] Support default port for IPv6 addresses (@clue)