+++ CHANGELOG.md
@@ -3,10 +3,10 @@ CHANGELOG
 
 ### 0.4.0 (2013-xx-xx)
 
+  * Feature: Added ChildProcess to run async child processes within the event loop (@jmikola)
   * BC break: [HttpClient] Drop unused `Response::getBody()`
   * BC break: Bump minimum PHP version to PHP 5.4, remove 5.3 specific hacks
   * Bug fix: [Dns] Properly resolve CNAME aliases
-
   * Bug fix: [EventLoop] No error on removing non-existent streams (@clue)
   * Bug fix: [EventLoop] Do not silently remove feof listeners in `LibEvLoop`
   * Bug fix: [Stream] Correctly detect closed connections