+++ CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGELOG
 ### 0.4.0 (2013-xx-xx)
 
   * Feature: Added ChildProcess to run async child processes within the event loop (@jmikola)
-  * BC break: [EventLoop] Added `EventLoopInterface::nextTick()`, implemented in all event loops (@jmalloc)
+  * BC break: [EventLoop] Added `EventLoopInterface::onNextTick()`, implemented in all event loops (@jmalloc)
   * BC break: [HttpClient] Drop unused `Response::getBody()`
   * BC break: Bump minimum PHP version to PHP 5.4, remove 5.3 specific hacks
   * Bug fix: [Dns] Properly resolve CNAME aliases